backend/requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
outcome==1.3.0.post0
pillow==10.3.0
pycparser==2.22
pydub==0.25.1
PySocks==1.7.1
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
